+++ src/hardware/cpu/cpu_comparison.cpp
@@ -19,7 +19,7 @@ bool cpu_all(Tensor *A){
     _profile(_CPU_ALL, 0);
 
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         if (A->ptr[i] != 1.0f){
             #pragma omp critical
             {
@@ -41,7 +41,7 @@ bool cpu_any(Tensor *A){
 
 
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         if (A->ptr[i] == 1.0f){
             #pragma omp critical
             {
@@ -60,7 +60,7 @@ bool cpu_any(Tensor *A){
 void cpu_isfinite(Tensor *A, Tensor* B){
     _profile(_CPU_ISFINITE, 0);
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         B->ptr[i] = std::isfinite(A->ptr[i]);
     }
     _profile(_CPU_ISFINITE, 1);
@@ -69,7 +69,7 @@ void cpu_isfinite(Tensor *A, Tensor* B){
 void cpu_isinf(Tensor *A, Tensor* B){
     _profile(_CPU_ISINF, 0);
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         B->ptr[i] = std::isinf(A->ptr[i]);
     }
     _profile(_CPU_ISINF, 1);
@@ -78,7 +78,7 @@ void cpu_isinf(Tensor *A, Tensor* B){
 void cpu_isnan(Tensor *A, Tensor* B){
     _profile(_CPU_ISNAN, 0);
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         B->ptr[i] = std::isnan(A->ptr[i]);
     }
     _profile(_CPU_ISNAN, 1);
@@ -88,7 +88,7 @@ void cpu_isnan(Tensor *A, Tensor* B){
 void cpu_isneginf(Tensor *A, Tensor* B){
     _profile(_CPU_ISNEGINF, 0);
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         B->ptr[i] = std::isinf(A->ptr[i]) && A->ptr[i] < 0.0f;
     }
     _profile(_CPU_ISNEGINF, 1);
@@ -97,7 +97,7 @@ void cpu_isneginf(Tensor *A, Tensor* B){
 void cpu_isposinf(Tensor *A, Tensor* B){
     _profile(_CPU_ISPOSINF, 0);
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         B->ptr[i] = std::isinf(A->ptr[i]) && A->ptr[i] > 0.0f;
     }
     _profile(_CPU_ISPOSINF, 1);
@@ -108,7 +108,7 @@ void cpu_isposinf(Tensor *A, Tensor* B){
 void cpu_logical_and(Tensor *A, Tensor *B, Tensor *C){
     _profile(_CPU_LOGICAL_AND, 0);
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         C->ptr[i] = (bool)A->ptr[i] & (bool)B->ptr[i];
     }
     _profile(_CPU_LOGICAL_AND, 1);
@@ -117,7 +117,7 @@ void cpu_logical_and(Tensor *A, Tensor *B, Tensor *C){
 void cpu_logical_or(Tensor *A, Tensor *B, Tensor *C){
     _profile(_CPU_LOGICAL_OR, 0);
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         C->ptr[i] = (bool)A->ptr[i] | (bool)B->ptr[i];
     }
     _profile(_CPU_LOGICAL_OR, 1);
@@ -126,7 +126,7 @@ void cpu_logical_or(Tensor *A, Tensor *B, Tensor *C){
 void cpu_logical_not(Tensor *A, Tensor *B){
     _profile(_CPU_LOGICAL_NOT, 0);
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         B->ptr[i] = !((bool)A->ptr[i]);  // why not use "~"
     }
     _profile(_CPU_LOGICAL_NOT, 1);
@@ -135,7 +135,7 @@ void cpu_logical_not(Tensor *A, Tensor *B){
 void cpu_logical_xor(Tensor *A, Tensor *B, Tensor *C){
     _profile(_CPU_LOGICAL_XOR, 0);
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         C->ptr[i] = (bool)A->ptr[i] ^ (bool)B->ptr[i];
     }
     _profile(_CPU_LOGICAL_XOR, 1);
@@ -149,7 +149,7 @@ bool cpu_allclose(Tensor *A, Tensor *B, float rtol, float atol, bool equal_nan){
 
     _profile(_CPU_ALLCLOSE, 0);
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         // Check if both values are NaN
         if (equal_nan && std::isnan(A->ptr[i]) && std::isnan(B->ptr[i])) {
             continue;
@@ -177,7 +177,7 @@ bool cpu_allclose_verbose(Tensor *A, Tensor *B, float rtol, float atol, bool equ
     bool allclose = true;
     int first_idx = -1;
 
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         // Check if both values are NaN
         if (equal_nan && std::isnan(A->ptr[i]) && std::isnan(B->ptr[i])) {
             continue;
@@ -199,7 +199,7 @@ bool cpu_allclose_verbose(Tensor *A, Tensor *B, float rtol, float atol, bool equ
 void cpu_isclose(Tensor *A, Tensor *B, Tensor *C, float rtol, float atol, bool equal_nan){
     _profile(_CPU_ISCLOSE, 0);
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         C->ptr[i] = ::fabsf(A->ptr[i] - B->ptr[i]) <= (atol + rtol * ::fabsf(B->ptr[i]));
     }
     _profile(_CPU_ISCLOSE, 1);
@@ -208,7 +208,7 @@ void cpu_isclose(Tensor *A, Tensor *B, Tensor *C, float rtol, float atol, bool e
 
 void cpu_greater(Tensor *A, Tensor *B, float v){
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         B->ptr[i] = A->ptr[i] > v;
     }
 }
@@ -217,7 +217,7 @@ void cpu_greater(Tensor *A, Tensor *B, Tensor *C){
     _profile(_CPU_GREATER, 0);
 
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         C->ptr[i] = A->ptr[i] > B->ptr[i];
     }
     _profile(_CPU_GREATER, 1);
@@ -226,7 +226,7 @@ void cpu_greater(Tensor *A, Tensor *B, Tensor *C){
 
 void cpu_greater_equal(Tensor *A, Tensor *B, float v){
 #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         B->ptr[i] = A->ptr[i] >= v;
     }
 }
@@ -235,7 +235,7 @@ void cpu_greater_equal(Tensor *A, Tensor *B, Tensor *C){
     _profile(_CPU_GREATER_EQUAL, 0);
 
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         C->ptr[i] = A->ptr[i] >= B->ptr[i];
     }
     _profile(_CPU_GREATER_EQUAL, 1);
@@ -243,7 +243,7 @@ void cpu_greater_equal(Tensor *A, Tensor *B, Tensor *C){
 
 void cpu_less(Tensor *A, Tensor *B, float v){
 #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         B->ptr[i] = A->ptr[i] < v;
     }
 }
@@ -252,7 +252,7 @@ void cpu_less(Tensor *A, Tensor *B, Tensor *C){
     _profile(_CPU_LESS, 0);
 
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         C->ptr[i] = A->ptr[i] < B->ptr[i];
     }
     _profile(_CPU_LESS, 1);
@@ -260,7 +260,7 @@ void cpu_less(Tensor *A, Tensor *B, Tensor *C){
 
 void cpu_less_equal(Tensor *A, Tensor *B, float v){
 #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         B->ptr[i] = A->ptr[i] <= v;
     }
 }
@@ -268,7 +268,7 @@ void cpu_less_equal(Tensor *A, Tensor *B, float v){
 void cpu_less_equal(Tensor *A, Tensor *B, Tensor *C){
     _profile(_CPU_LESS_EQUAL, 0);
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         C->ptr[i] = A->ptr[i] <= B->ptr[i];
     }
     _profile(_CPU_LESS_EQUAL, 1);
@@ -276,7 +276,7 @@ void cpu_less_equal(Tensor *A, Tensor *B, Tensor *C){
 
 void cpu_equal(Tensor *A, Tensor *B, float v){
 #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         B->ptr[i] = A->ptr[i] == v;
     }
 }
@@ -284,7 +284,7 @@ void cpu_equal(Tensor *A, Tensor *B, float v){
 void cpu_equal(Tensor *A, Tensor *B, Tensor *C){
     _profile(_CPU_EQUAL, 0);
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         C->ptr[i] = A->ptr[i] == B->ptr[i];
     }
     _profile(_CPU_EQUAL, 1);
@@ -292,7 +292,7 @@ void cpu_equal(Tensor *A, Tensor *B, Tensor *C){
 
 void cpu_not_equal(Tensor *A, Tensor *B, float v){
 #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         B->ptr[i] = A->ptr[i] != v;
     }
 }
@@ -300,7 +300,7 @@ void cpu_not_equal(Tensor *A, Tensor *B, float v){
 void cpu_not_equal(Tensor *A, Tensor *B, Tensor *C){
     _profile(_CPU_NOT_EQUAL, 0);
     #pragma omp parallel for
-    for (int i = 0; i < A->size; ++i){
+    for (unsigned long int i = 0; i < A->size; ++i){
         C->ptr[i] = A->ptr[i] != B->ptr[i];
     }
     _profile(_CPU_NOT_EQUAL, 1);
